fix: Refactor the module structure by adding model, service, and utility modules.
