+++ watchman-backend/src/model/service.rs
@@ -4,7 +4,7 @@ use diesel::{
 };
 use serde::{Deserialize, Serialize};
 
-use crate::models::schema::service_table::{self, dsl::*};
+use crate::model::schema::service_table::{self, dsl::*};
 
 static NOT_FOUND_CODE: u8 = 1;
 static TMI_ERROR_CODE: u8 = 2;
